requirements.txt
fastapi
uvicorn
python-multipart
pyyaml
pydantic
numpy
pandas
scipy
h5py
matplotlib
plotly
kaleido
torch
torchvision
snntorch
tqdm
loguru
websockets
flask
flask-cors
click
